moved sortable to components; removed interventions
