--- client/src/containers/interventions/smart-filters/materials/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
