--- client/src/containers/interventions/smart-filters/origin-regions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
