--- client/src/containers/interventions/smart-filters/suppliers/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
